refact: modularizing user component
