+++ src/index.css
@@ -163,11 +163,14 @@ html {
 }
 
 .container__user {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
   overflow-y: hidden;
   margin-top: 75px;
   background-color: #111111;
+}
+
+.container__grid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
 }
 
 .container__playlist,
@@ -262,9 +265,21 @@ a {
   justify-content: center;
 }
 
+#table__row {
+  justify-content: space-between;
+  background: #111111;
+  border-radius: 5px;
+  margin-top: 2em;
+}
+
+#table__row:hover {
+  background: #1ed760;
+  color: black;
+  cursor: pointer;
+}
+
 td {
-  border-right: 1px solid;
-  border-bottom: 1px solid;
+  padding-top: 5px;
 }
 
 
@@ -279,6 +294,7 @@ tr:nth-child(n + 4) {
 #table__head {
   font-weight: bolder;
   font-size: large;
+  border-bottom: 1px solid white;
 }
 
 .arrow {
@@ -322,12 +338,14 @@ ul {
   flex-direction: row;
   justify-content: flex-end;
   color: #f7f7f7;
-  border-bottom: 1px solid #f7f7f7;
-  border-left: 1px solid #f7f7f7;
-  border-right: 1px solid #f7f7f7;
-  border-collapse: collapse;
+  box-shadow: 0px 4px 0px 2px rgba(255, 255, 255, 0.918);
+
   padding-top: 10px;
   padding-bottom: 10px;
-  padding-right: 2em;
-  margin-right: 5px;
+  padding-right: 1em;
+}
+
+.sidebar__loading {
+  display: flex;
+  align-items: center;
 }
